+++ MiraAPI/Patches/Keybinds/KeybindsPatch.cs
@@ -1,5 +1,6 @@
 using HarmonyLib;
 using MiraAPI.Keybinds;
+using MiraAPI.Utilities;
 using Rewired;
 
 namespace MiraAPI.Patches.Keybinds
