+++ MiraAPI/Utilities/Extensions.cs
@@ -10,6 +10,8 @@ using MiraAPI.Networking;
 using MiraAPI.Roles;
 using MiraAPI.Voting;
 using Reactor.Utilities;
+using Rewired;
+using Rewired.Data;
 using TMPro;
 using UnityEngine;
 
@@ -92,7 +94,7 @@ public static class Extensions
     /// <param name="timer">The current timer value.</param>
     /// <param name="maxTimer">The maximum timer value.</param>
     /// <param name="format">The format string to use for the timer text.</param>
-    public static void SetCooldownFormat(this ActionButton button, float timer, float maxTimer, string format="0")
+    public static void SetCooldownFormat(this ActionButton button, float timer, float maxTimer, string format = "0")
     {
         var num = Mathf.Clamp(timer / maxTimer, 0f, 1f);
         button.isCoolingDown = num > 0f;
@@ -113,7 +115,7 @@ public static class Extensions
     /// <param name="timer">The current timer value.</param>
     /// <param name="maxTimer">The maximum timer value.</param>
     /// <param name="format">The format string to use for the timer text.</param>
-    public static void SetFillUpFormat(this ActionButton button, float timer, float maxTimer, string format="0")
+    public static void SetFillUpFormat(this ActionButton button, float timer, float maxTimer, string format = "0")
     {
         var num = Mathf.Clamp(timer / maxTimer, 0f, 1f);
         button.isCoolingDown = num > 0f;
@@ -543,5 +545,45 @@ public static class Extensions
         renderer.material.SetFloat(ShaderID.Outline, color.HasValue ? 1 : 0);
         renderer.material.SetColor(ShaderID.OutlineColor, color ?? Color.clear);
         renderer.material.SetColor(ShaderID.AddColor, color ?? Color.clear);
+    }
+
+    /// <summary>
+    /// Registers a new mod keybind as a user-assignable button action in Rewired.
+    /// </summary>
+    /// <param name="userData">The Rewired user data to add the action to.</param>
+    /// <param name="actionName">The internal name of the action.</param>
+    /// <param name="description">Text shown in the rebinding UI.</param>
+    /// <param name="key">The default key to assign to this action.</param>
+    /// <param name="category">Category ID to group actions in Rewired (default is 0).</param>
+    /// <param name="elementIdentifierId">The element identifier ID (default is -1, meaning none specified).</param>
+    /// <param name="type">The <see cref="InputActionType"/> for this action (default is Button).</param>
+    /// <returns>The action ID of the newly registered action.</returns>
+    public static int RegisterModBind(this UserData userData, string actionName, string description, KeyboardKeyCode key, int category = 0, int elementIdentifierId = -1, InputActionType type = InputActionType.Button)
+    {
+        userData.AddAction(category);
+
+        var action = userData.GetAction(userData.actions.Count - 1)!;
+
+        action.name = actionName;
+        action.descriptiveName = description;
+        action.categoryId = category;
+        action.type = type;
+        action.userAssignable = true;
+
+        var map = new ActionElementMap
+        {
+            _elementIdentifierId = elementIdentifierId,
+            _actionId = action.id,
+            _elementType = ControllerElementType.Button,
+            _axisContribution = Pole.Positive,
+            _keyboardKeyCode = key,
+            _modifierKey1 = ModifierKey.None,
+            _modifierKey2 = ModifierKey.None,
+            _modifierKey3 = ModifierKey.None,
+        };
+        userData.keyboardMaps[0].actionElementMaps.Add(map);
+        userData.joystickMaps[0].actionElementMaps.Add(map);
+
+        return action.id;
     }
 }
